Library is generated and parses correctly
