+++ package.json
@@ -14,10 +14,7 @@
     "verbose": true,
     "bail": false,
     "collectCoverage": false,
-    "modulePathIgnorePatterns": [
-      "node_modules",
-      "api.test.js"
-    ]
+    "modulePathIgnorePatterns": ["node_modules", "api.test.js", "sandbox"]
   },
   "dependencies": {
     "chalk": "^2.3.0",
